writed in one func

+++ module_2/hw-4/sort_in_threads.py
@@ -3,77 +3,35 @@ from pathlib import Path
 import shutil
 from threading import Thread
 from concurrent.futures import ThreadPoolExecutor
-
 import scan
 from normalize import normalize
 
 
-def handle_image(root_folder: Path, dist: str):
-    for file in scan.IMAGE:
+def handle_files(root_folder: Path, folder_for_sort, dist: str):
+    for file in folder_for_sort:
         target_folder = root_folder / dist
         target_folder.mkdir(exist_ok=True)
         ext = Path(file).suffix
         new_name = normalize(file.name.replace(ext, "")) + ext
-        file.replace(target_folder / new_name)
 
-
-def handle_video(root_folder: Path, dist: str):
-    for file in scan.VIDEO:
-        target_folder = root_folder / dist
-        target_folder.mkdir(exist_ok=True)
-        ext = Path(file).suffix
-        new_name = normalize(file.name.replace(ext, "")) + ext
-        file.replace(target_folder / new_name)
-
-
-def handle_audio(root_folder: Path, dist: str):
-    for file in scan.AUDIO:
-        target_folder = root_folder / dist
-        target_folder.mkdir(exist_ok=True)
-        ext = Path(file).suffix
-        new_name = normalize(file.name.replace(ext, "")) + ext
-        file.replace(target_folder / new_name)
-
-
-def handle_doc(root_folder: Path, dist: str):
-    for file in scan.DOC:
-        target_folder = root_folder / dist
-        target_folder.mkdir(exist_ok=True)
-        ext = Path(file).suffix
-        new_name = normalize(file.name.replace(ext, "")) + ext
-        file.replace(target_folder / new_name)
-
-
-def handle_other(root_folder: Path, dist: str):
-    for file in scan.OTHER:
-        target_folder = root_folder / dist
-        target_folder.mkdir(exist_ok=True)
-        ext = Path(file).suffix
-        new_name = normalize(file.name.replace(ext, "")) + ext
-        file.replace(target_folder / new_name)
-
-
-def handle_archive(root_folder: Path, dist):
-    for file in scan.ARCH:
-        target_folder = root_folder / dist
-        target_folder.mkdir(exist_ok=True)
-        ext = Path(file).suffix
-        folder_for_archive = normalize(file.name.replace(ext, ""))
-        archive_folder = target_folder / folder_for_archive
-        archive_folder.mkdir(exist_ok=True)
-        try:
-            shutil.unpack_archive(str(file.resolve()), str(archive_folder.resolve()))
-        except shutil.ReadError:
-            archive_folder.rmdir()
-            return
-        file.unlink()
+        if dist != 'archives':
+            file.replace(target_folder / new_name)
+        else:
+            archive_folder = target_folder / new_name
+            archive_folder.mkdir(exist_ok=True)
+            try:
+                shutil.unpack_archive(str(file.resolve()), str(archive_folder.resolve()))
+            except shutil.ReadError:
+                archive_folder.rmdir()
+                return
+            file.unlink()
 
 
 def handle_folder(folder: Path):
     try:
         folder.rmdir()
     except OSError:
-        pass
+        print("System error")
 
 
 workers = len(scan.REGISTERED_EXTENSIONS.keys())
@@ -85,12 +43,12 @@ def main(folder):
     thr.join()
 
     with ThreadPoolExecutor(max_workers=workers) as executor:
-        executor.submit(handle_image, folder, 'images')
-        executor.submit(handle_video, folder, 'video')
-        executor.submit(handle_audio, folder, 'audio')
-        executor.submit(handle_doc, folder, 'documents')
-        executor.submit(handle_other, folder, 'other')
-        executor.submit(handle_archive, folder, 'archives')
+        executor.submit(handle_files, folder, scan.IMAGE, 'images')
+        executor.submit(handle_files, folder, scan.VIDEO, 'video')
+        executor.submit(handle_files, folder, scan.AUDIO, 'audio')
+        executor.submit(handle_files, folder, scan.DOC, 'documents')
+        executor.submit(handle_files, folder, scan.OTHER, folder, 'other')
+        executor.submit(handle_files, folder, scan.ARCH, 'archives')
 
     for item in scan.FOLDERS:
         handle_folder(item)
@@ -103,4 +61,3 @@ if __name__ == "__main__":
     sort_folder = Path(scan_path)
 
     main(sort_folder.resolve())
-
